Introduced IndexingScheme and IndexStructure classes, which keep track of all the different index structures (names of vars and kinds of refs) that make up the complete indexing-scheme for the tracematch.
Integrated the existing code-generation with these new helper-classes.

Wrote code to generate the new or() method for indexed constraints,
which takes a collection of disjuncts and inserts them all into
the current indexing tree. It uses a smart lookup-algorithm which
avoids looking up keys in the tree twice when consecutive disjuncts
share some variable bindings.