start the shadow *after* the new-statement, i.e. right before the constructor call, not before the new-statement;
This is for increased precision of must-alias analysis. Consider this case:

x=null;

label0:
 x = new O;
 x.<init>();

label1:
//after throwing

label2:
...
trap: from label0 to label1 with label1

Here at label2, x will have two possible definitions, due to the trap. If we move label0 one statement further down, the definition at label2 is unique.

Since a new-statement can only fail with an OutOfMemoryError and in this case the program will fail anyway, it should be sound to have the trap start after it.