Implementation of flow-sensitive optimizations for dependent advice
Dependent advice now support a state-machine based syntax that basically allows a programmer or code generation tool to annotate an aspect with a state machine which describes the aspects expected order or advice execution. abc can then exploit these annotations to optimize the aspect, even flow-sensitively, just as was previously the case for tracmatches.

The basic entity which all static whole-program optimizations now operate on is a TracePattern. A TracePattern is basically an abstraction of a TraceMatch. It combines a SimpleStateMachine with a set of symbols and other auxiliary information. SimpleStateMachine is an abstraction of TMStateMachine. TMStateMachine was changed so that it extends SimpleStateMachine, which now holds basic funcitonality common to state machines in general, such as creating transitions and determinization, etc. This allows abc.da to re-use this functionality even for state machines that do not come from a tracematch declaration.

The entire flow-sensitive analysis was moved from abc.tmwpopt to abc.da. All that abc.tmwpopt now does is creating a TracePattern from a TraceMatch and registering that with abc.da.

The entire flow-sensitive analysis now stores the shadow history in a more sensible way. This allowed us to remove some minor bugs w.r.t. history storage.